Include Scripts directory in the Windows package by adding .keep file.
Currently conda will fail to create an entry_point script/binary for
downstream noarch: python packages unless the parent directory
(%PREFIX%\Scripts) already exists. This ensures that the directory will
always be created when pypy is installed, which happens to be already
true for the cpython packages. This is only an issue on Windows because
non-Windows entry_points go to $PREFIX/bin which exists because other
files already go there.